# api/openapi.yaml
openapi: 3.0.3
info:
  title: Licenses Admin API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /products:
    get:
      summary: List products
      parameters:
        - in: query
          name: manufacturerId
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProduct'
    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{id}:
    get:
      summary: Get product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found
    put:
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found
    delete:
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: No content
  /products/{productId}/variants:
    get:
      summary: List variants for product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProductVariant'
    post:
      summary: Create product variant
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductVariantCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVariant'
...
  /manufacturers:
    get:
      summary: List manufacturers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageManufacturer'
    post:
      summary: Create manufacturer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
  /manufacturers/{id}:
    get:
      summary: Get manufacturer by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
        '404':
          description: Not found
    put:
      summary: Update manufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
        '404':
          description: Not found
    delete:
      summary: Delete manufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: No content
components:
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
  schemas:
    UUID:
      type: string
      format: uuid
...
    Product:
      type: object
      required: [id, key, name]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        key: { type: string }
        name: { type: string }
        description: { type: string }
        manufacturerId:
          type: string
          format: uuid
          nullable: true
    ProductCreateRequest:
      type: object
      required: [key, name]
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string }
        manufacturerId:
          type: string
          format: uuid
          nullable: true
    ProductUpdateRequest:
      type: object
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string }
        manufacturerId:
          type: string
          format: uuid
          nullable: true
    PageProduct:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Product' }
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
    ProductVariant:
      type: object
      required: [id, productId, key, name]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        productId: { $ref: '#/components/schemas/UUID' }
        key: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price:
          type: number
          format: double
          nullable: true
    ProductVariantCreateRequest:
      type: object
      required: [key, name]
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price:
          type: number
          format: double
          nullable: true
...
    PageProductVariant:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/ProductVariant' }
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
...
    Manufacturer:
      type: object
      required: [id, name]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        url: { type: string, nullable: true }
        description: { type: string, nullable: true }
    ManufacturerCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        url: { type: string, nullable: true }
        description: { type: string, nullable: true }
    ManufacturerUpdateRequest:
      type: object
      properties:
        name: { type: string }
        url: { type: string, nullable: true }
        description: { type: string, nullable: true }
    PageManufacturer:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Manufacturer' }
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }