# target/generated-sources/openapi/src/main/java/io/meimberg/licenses/web/api/ManufacturersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.meimberg.licenses.web.api;

import io.meimberg.licenses.web.dto.Manufacturer;
import io.meimberg.licenses.web.dto.ManufacturerCreateRequest;
import io.meimberg.licenses.web.dto.ManufacturerUpdateRequest;
import io.meimberg.licenses.web.dto.PageManufacturer;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-23T15:50:51.445317955+01:00[Europe/Berlin]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "manufacturers", description = "the manufacturers API")
public interface ManufacturersApi {

    /**
     * GET /manufacturers : List manufacturers
     *
     * @param page  (optional)
     * @param size  (optional)
     * @param sort  (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "manufacturersGet",
        summary = "List manufacturers",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageManufacturer.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/manufacturers",
        produces = { "application/json" }
    )
    
    ResponseEntity<PageManufacturer> manufacturersGet(
        @Min(0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) Integer page,
        @Min(1) @Max(200) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) Integer size,
        @Parameter(name = "sort", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) String sort
    );


    /**
     * DELETE /manufacturers/{id} : Delete manufacturer
     *
     * @param id  (required)
     * @return No content (status code 204)
     */
    @Operation(
        operationId = "manufacturersIdDelete",
        summary = "Delete manufacturer",
        responses = {
            @ApiResponse(responseCode = "204", description = "No content")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/manufacturers/{id}"
    )
    
    ResponseEntity<Void> manufacturersIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /manufacturers/{id} : Get manufacturer by id
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "manufacturersIdGet",
        summary = "Get manufacturer by id",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Manufacturer.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/manufacturers/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Manufacturer> manufacturersIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * PUT /manufacturers/{id} : Update manufacturer
     *
     * @param id  (required)
     * @param manufacturerUpdateRequest  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "manufacturersIdPut",
        summary = "Update manufacturer",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Manufacturer.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/manufacturers/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Manufacturer> manufacturersIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "ManufacturerUpdateRequest", description = "", required = true) @Valid @RequestBody ManufacturerUpdateRequest manufacturerUpdateRequest
    );


    /**
     * POST /manufacturers : Create manufacturer
     *
     * @param manufacturerCreateRequest  (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "manufacturersPost",
        summary = "Create manufacturer",
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Manufacturer.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/manufacturers",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Manufacturer> manufacturersPost(
        @Parameter(name = "ManufacturerCreateRequest", description = "", required = true) @Valid @RequestBody ManufacturerCreateRequest manufacturerCreateRequest
    );

}