# target/generated-sources/openapi/src/main/java/io/meimberg/licenses/web/api/ProductsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.meimberg.licenses.web.api;

import io.meimberg.licenses.web.dto.PageProduct;
import io.meimberg.licenses.web.dto.PageProductVariant;
import io.meimberg.licenses.web.dto.Product;
import io.meimberg.licenses.web.dto.ProductCreateRequest;
import io.meimberg.licenses.web.dto.ProductUpdateRequest;
import io.meimberg.licenses.web.dto.ProductVariant;
import io.meimberg.licenses.web.dto.ProductVariantCreateRequest;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-23T15:50:51.445317955+01:00[Europe/Berlin]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "products", description = "the products API")
public interface ProductsApi {

    /**
     * GET /products : List products
     *
     * @param manufacturerId  (optional)
     * @param page  (optional)
     * @param size  (optional)
     * @param sort  (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "productsGet",
        summary = "List products",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageProduct.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products",
        produces = { "application/json" }
    )
    
    ResponseEntity<PageProduct> productsGet(
        @Parameter(name = "manufacturerId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "manufacturerId", required = false) UUID manufacturerId,
        @Min(0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) Integer page,
        @Min(1) @Max(200) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) Integer size,
        @Parameter(name = "sort", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) String sort
    );


    /**
     * DELETE /products/{id} : Delete product
     *
     * @param id  (required)
     * @return No content (status code 204)
     */
    @Operation(
        operationId = "productsIdDelete",
        summary = "Delete product",
        responses = {
            @ApiResponse(responseCode = "204", description = "No content")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/products/{id}"
    )
    
    ResponseEntity<Void> productsIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /products/{id} : Get product by id
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "productsIdGet",
        summary = "Get product by id",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Product> productsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * PUT /products/{id} : Update product
     *
     * @param id  (required)
     * @param productUpdateRequest  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "productsIdPut",
        summary = "Update product",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/products/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Product> productsIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "ProductUpdateRequest", description = "", required = true) @Valid @RequestBody ProductUpdateRequest productUpdateRequest
    );


    /**
     * POST /products : Create product
     *
     * @param productCreateRequest  (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "productsPost",
        summary = "Create product",
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Product> productsPost(
        @Parameter(name = "ProductCreateRequest", description = "", required = true) @Valid @RequestBody ProductCreateRequest productCreateRequest
    );


    /**
     * GET /products/{productId}/variants : List variants for product
     *
     * @param productId  (required)
     * @param page  (optional)
     * @param size  (optional)
     * @param sort  (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "productsProductIdVariantsGet",
        summary = "List variants for product",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageProductVariant.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/{productId}/variants",
        produces = { "application/json" }
    )
    
    ResponseEntity<PageProductVariant> productsProductIdVariantsGet(
        @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId,
        @Min(0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) Integer page,
        @Min(1) @Max(200) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) Integer size,
        @Parameter(name = "sort", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) String sort
    );


    /**
     * POST /products/{productId}/variants : Create product variant
     *
     * @param productId  (required)
     * @param productVariantCreateRequest  (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "productsProductIdVariantsPost",
        summary = "Create product variant",
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductVariant.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/products/{productId}/variants",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProductVariant> productsProductIdVariantsPost(
        @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId,
        @Parameter(name = "ProductVariantCreateRequest", description = "", required = true) @Valid @RequestBody ProductVariantCreateRequest productVariantCreateRequest
    );

}